Move the swipeable implementation to the "events" folder
