+++ src/components/canvas/canvas.tsx
@@ -15,7 +15,7 @@ import {
   DisableableComponent,
   VisibilityComponent
 } from "../common/interfaces";
-import { Swipeable, makeSwipeable } from "../common/swipeable";
+import { Swipeable, makeSwipeable } from "../common/events/swipeable";
 
 // Class transforms
 import { getClassesWithoutFocus } from "../common/css-transforms/css-transforms";
